+++ lib/src/commands/list/organizations.dart
@@ -4,13 +4,11 @@
 // Use of this source code is governed by terms that can be
 // found in the LICENSE file in the root of this package.
 
-import 'dart:io';
-
 import 'package:args/command_runner.dart';
 import 'package:gg_console_colors/gg_console_colors.dart';
 import 'package:gg_log/gg_log.dart';
 import '../../backend/list_backend.dart';
-import 'package:path/path.dart' as path;
+import '../../backend/workspace_utils.dart';
 
 /// Command to list all organizations from repos in the master workspace.
 class ListOrganizationsCommand extends Command<dynamic> {
@@ -19,8 +17,8 @@ class ListOrganizationsCommand extends Command<dynamic> {
     required this.ggLog,
     String? workspacePath,
     // coverage:ignore-start
-  }) : workspacePath = workspacePath ??
-            path.join(Directory.current.path, 'kidney_ws_master');
+  }) : workspacePath =
+            workspacePath ?? WorkspaceUtils.defaultMasterWorkspacePath();
   // coverage:ignore-end
 
   /// The log function.