+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Ignore dart dependencies in add-deps
 - Log messages start with upper case
 - Update Readme
+- execute list repos, list organizations and create ticket always in kidney workspace
 
 ### Removed